+++ src/main/java/be/dash/dashserver/api/core/auth/dto/LoginResponse.java
@@ -1,6 +1,6 @@
 package be.dash.dashserver.api.core.auth.dto;
 
-import be.dash.dashserver.core.LoginResult;
+import be.dash.dashserver.core.auth.dto.LoginResult;
 
 public record LoginResponse(
         String accessToken,
