+++ src/main/java/be/dash/dashserver/external/config/s3/S3Properties.java
@@ -1,4 +1,4 @@
-package be.dash.dashserver.external.config;
+package be.dash.dashserver.external.config.s3;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
